Add a local dev environment based on docker-compose
The environment can be accessed by running "docker-compose up" in the
project's root directory.

The environment attempts to mimic the production envirmonment.

In the local dev environment there is a single nginx container that
acts as a reverse proxy for three node containers running the app. The
node containers connect to a single redis container.

In the production environment three pods, each comprising of a nginx
container and a node container, connect to a single redis pod.