Refactor AJAX controller code
Refactor the new AJAX controller code to enable a wider variety of
uses. AjaxController should be used for the standard controller on
forms which require AJAX submits. AjaxEditController should be used on
forms which are used for editing previously entered data. The
difference being that on 'edit' forms we do not require all fields to be
present, only the ones we are updating. This is because we already have
the data we require.

This will allow us to safely use the AjaxController to submit complete
forms. e.g. Adding a new 'report' in ROTM.