Use valid SPDX licence expression
The Software Package Data Exchange defines a standard format for
machine-readable licence data. Previously, we were using a non-standard
identifier for the GPLv2; this commit amends it to adhere to the
standard.

NPM complains about the invalid SPDX expression when building the app;
by fixing this we get rid of the warning.