[RTMD-141] - New page: /image (#117)
* RTMD-141 New page: /image

file-input functionality for images
Hides and shows buttons & file inputs based on JS being enabled or disabled

* Add skip-step behaviour for /image & test setup

This will ensure that the /image step is complete so that a user can navigate to the next step by clicking on a link

* Add forking condition for /image link

previously the user would be navigated to the /add-image step if they clicked on the link rather than /check-your-report

* Amend skipstep behaviour getValues

utilise values in getValues
amend also the tests as a result

* Amend skip-step behaviour to take steps as arguments

Previously a fork & behaviour was used to skip a step.  This has been removed.  It makes it easier to understand if it is all in a behaviour
Amends the tests as well

* Simplify skip-step behaviour & add prereq to /check-your-report

The prereq is needed so that the link in /image allows the user to jump to /check-your-report