Generate image thumbnail when saving uploaded image (#124)
* Generate image thumbnail when saving uploaded image

Uses `sharp` to create a thumbnail of the uploaded file and saves the base64 string to the session.

* Unset image properties from model when uploading a new image

So if the process fails at any point artefacts are not left in place.

* Simplify image saving process

Remove the additional properties and structure which are not necessary for validating and processing the image

* Swap `sharp` for `jimp` for image thumbnailing

Jimp doesn't have any external dependencies so will install on our docker containers. Sharp needs `make` to install binaries, so doesn't instll cleanly.