Don't use ValidationError for file-vault failures (#123)
When validation errors are thrown they display with a message "Please fix the following error". If saving the file to file vault has failed, then there is nothing the user can "fix" to make it work, so displaying a validation message is not appropriate, and an application error should be displayed.