Run acceptance tests in headless chrome (#121)
* Use headless chrome for running acceptance tests in CI

* Add image upload test

* Orchestrate chromedriver for running tests locally

Wrapping the test command in `funkie` with chromedriver sves a local user from needing to run chromedriver independently. This is not necessary in CI since there is a pre-existing chromedriver service running in Drone.

* Simplify happy path tests

It's no longer necessary to divert around image upload for the happy path, we can happily upload an image and follow that journey.

* Remove workaround for phantom bug

The random extra click step was only to work around a bug in phantom. Since there's no phantom anymore then there's no need for the workaround.

* Fix acceptance test field names

* Ensure happy path test accepts the image that it uploads

Otherwise it gets stuck in an infinite loop around re-uploading the same image.

* If configured to use mocks then automatically use mock urls

Use a single piece of configuration to enable the use of mocks rather than needing to set `USE_MOCKS` *and* unset `FILE_VAULT_URL`

* Add re-usable default form values for acceptance tests