Add uploaded image preview to confirmation page
Shows a thumbnail fo the iamge a user uploaded when they reach the check answers page.

Content is hard coded into template to avoid issues with escaping user input when rendering as html - if it was rendered in the default way we would have to show all values as unescaped html, which would introduce an XSS attack vector.