Allow confirming images in an "edit" journey
This allows the submission ofa new image to go to the yes/no confirm page, and similarly, the selecting "no" to return to the upload page.

This is a mild hack because in theory the "continue on edit" functionality should only apply to the "no" fork, and selecting "yes" should return to the confirm page, but since that's the next page anyway, it sort of works.