[Rtmd 146] - Caseworker email (#128)
* [RTMD-146] - Send caseworker email upon submission

Send an email to the configured caseworker email with a link to the s3 image

* Disable sending email in CI environments

Use stub email transport, because using a real transport fails.

* Set caseworker email config in drone builds

If this is not set then acceptance tests error out

* Casworker email behaviour- add unique id

A unique reference number is generated with a timestamp and displayed on the caseworker email

* optional config for Email transport

This is to allow development with smtp rather than ses.  Smtp requires some other things to be set

* Remove image label translation as this is not being used.

The image label is in the caseworker.html and not in the translation.  It is in the html because it requires a link with an achor.

* Remove unnecessary environment variable from CI test

The `CASEWORKER_EMAIL` environment variable is not required in order to run acceptance tests. Remove it.

* Simplify email data structure

There was no need for the data passed to the email to contain the additional structure of `{ data: [ ... ] }` where it could simply pass the table and the image url as top level properties.

* Ensure value is returned from arrow function

Previous version had an arrow function with curly braces but not `return` so email subject was being sent as undefined. Remove the curlies so that email subject is returned from the function correctly.

* Updated caseworker email content

Following conversation with Nadia, the main heading has been removed - to avoid duplicate headings with the subject line also included in the email body - and the subject line has been updated.